+++ packages/orq-rc/src/models/components/partdoneevent.ts
@@ -79,7 +79,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kf17698k1zt1camx5a7q1b0d"),
+  _id: z.string().default("reasoning_01kf2rj5qpg4cvjttscwdtp5sq"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -104,7 +104,7 @@ export const PartReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   PartReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kf17698k1zt1camx5a7q1b0d"),
+  id: z.string().default("reasoning_01kf2rj5qpg4cvjttscwdtp5sq"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$outboundSchema,
   reasoning: z.string(),