+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -157,7 +157,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-01-15T16:17:55.965Z",
+    "2026-01-16T06:40:45.388Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -190,7 +190,7 @@ export const FileListData$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2026-01-15T16:17:55.965Z"))
+  created: z.date().default(() => new Date("2026-01-16T06:40:45.388Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {