+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -211,7 +211,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-15T16:17:52.624Z",
+    "2026-01-16T06:40:41.805Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -250,7 +250,7 @@ export const ListDatasetsData$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2026-01-15T16:17:52.624Z"))
+  updated: z.date().default(() => new Date("2026-01-16T06:40:41.805Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {